requirements.txt
mangum
fastapi
psycopg2-binary
python-multipart
uvicorn
boto3